+++ src/leiningen/new/db.clj
@@ -12,7 +12,7 @@
 
 (defn db-dependencies [options]
   (into [['luminus-migrations "0.7.1"]
-         ['conman "0.9.0"]]
+         ['conman "0.9.1"]]
         ({:postgres [['org.postgresql/postgresql "42.2.18"]]
           :mysql    [['mysql/mysql-connector-java "8.0.18"]
                      ['com.google.protobuf/protobuf-java "3.8.0"]]