+++ src/leiningen/new/re_frame.clj
@@ -11,7 +11,8 @@
            re-frame-assets)
      (-> options
          (assoc :re-frame true)
-         (append-options :dependencies [['re-frame "1.1.1"]
-                                        ['day8.re-frame/http-fx "0.2.1"]])
-         (append-options :dev-dependencies [['day8.re-frame/re-frame-10x "0.7.0"]]))]
+         (append-options :dependencies [['re-frame "1.1.2"]
+                                        ['day8.re-frame/http-fx "0.2.2"]])
+         (append-options :dev-dependencies [['re-frisk "1.3.5"]
+                                            #_['day8.re-frame/re-frame-10x "0.7.0"]]))]
     state))