+++ src/leiningen/new/reitit.clj
@@ -10,7 +10,7 @@
     [(into (remove-conflicting-assets assets "home.clj" "docs.md") reitit-assets)
      (-> options
          (append-options :dependencies [['metosin/reitit "0.5.10"]
-                                        ['metosin/jsonista "0.2.7"]
-                                        ['com.fasterxml.jackson.core/jackson-core "2.11.3"]])
+                                        #_['metosin/jsonista "0.2.7"]
+                                        #_['com.fasterxml.jackson.core/jackson-core "2.12.0"]])
          (assoc :reitit true))]
     state))