+++ src/leiningen/new/shadow_cljs.clj
@@ -28,7 +28,7 @@
                          {:entries [(project-ns-symbol project-ns ".app")]}}
             :devtools   (merge {:watch-dir "resources/public"}
                                (when (some #{"+re-frame"} features)
-                                 {:preloads ['day8.re-frame-10x.preload]}))}
+                                 {:preloads ['re-frisk.preload #_'day8.re-frame-10x.preload]}))}
            (when (some #{"+re-frame"} features)
              {:dev {:closure-defines {"re_frame.trace.trace_enabled_QMARK_" true}}}))